refactor(team): extract auth + pub/sub fanout from broker.go
Phase 3 second slice. Two small cohesive extractions:

(1) broker_auth.go (62 LOC, new): auth surface
    - generateToken: cryptographically random shared secret
    - requireAuth: Bearer-token middleware (also accepts ?token=
      query param for EventSource which can't set headers)
    - handleWebToken: localhost-only token endpoint (gated by
      RemoteAddr + Host loopback check to defeat DNS rebinding)

(2) broker_publish.go (49 LOC, new): pub/sub fanout
    - appendMessageLocked: persist + publish in one step
    - publishMessageLocked / publishActionLocked /
      publishActivityLocked: the three "always-on" channels with
      non-blocking sends (default-on-full so a slow subscriber
      can't stall the publisher)

  - internal/team/broker.go: -59 LOC (2859 -> 2800)
  - internal/team/broker_auth.go: +62 LOC (new)
  - internal/team/broker_publish.go: +49 LOC (new)

Cumulative broker.go reduction across Phase 3:
  3404 -> 2859 (web proxy + sse)         -545
       -> 2800 (auth + publish)          -59
  Net: -604 LOC; broker.go shrunk by 18%.

No behavior change. internal/team tests pass with -race.
