+++ internal/team/broker_auth.go
@@ -34,16 +34,31 @@ package team
 //     handleEvents calls requestHasBrokerAuth at the top.
 
 import (
+	"crypto/rand"
+	"encoding/hex"
+	"encoding/json"
+	"fmt"
 	"io"
 	"net/http"
+	"time"
 )
+
+// generateToken returns a cryptographically random hex token.
+func generateToken() string {
+	b := make([]byte, 16)
+	if _, err := rand.Read(b); err != nil {
+		// Fallback: this should never happen on modern systems.
+		return fmt.Sprintf("wuphf-%d", time.Now().UnixNano())
+	}
+	return hex.EncodeToString(b)
+}
 
 // withAuth wraps a handler to require Bearer-token authentication.
 //
 // This is the canonical middleware for protected broker routes. The older
-// requireAuth method (above on Broker) is kept as a thin alias so that the
-// existing 100+ route registrations don't all churn in this PR; new routes
-// should call b.withAuth directly.
+// requireAuth method is kept as a thin alias so that the existing 100+ route
+// registrations don't all churn in this PR; new routes should call b.withAuth
+// directly.
 //
 // Both names share an implementation, so the auth-route assertion in
 // broker_workspaces_test.go covers both call sites.
@@ -59,4 +74,28 @@ func (b *Broker) withAuth(next http.HandlerFunc) http.HandlerFunc {
 		w.WriteHeader(http.StatusUnauthorized)
 		_, _ = io.WriteString(w, `{"error":"unauthorized"}`)
 	}
+}
+
+// requireAuth wraps a handler to enforce Bearer token authentication.
+// Accepts token via Authorization header or ?token= query parameter for
+// EventSource, which can't set headers.
+func (b *Broker) requireAuth(next http.HandlerFunc) http.HandlerFunc {
+	return b.withAuth(next)
+}
+
+// handleWebToken returns the broker token to localhost clients without requiring auth.
+// This lets the web UI fetch the token to authenticate subsequent API calls.
+//
+// DNS rebinding: even though the listener binds 127.0.0.1, an attacker's
+// DNS record with a short TTL can point rebind.example.com at 127.0.0.1
+// after the browser's origin check passes. Go's default mux routes purely
+// on path, so without an explicit Host check the response would flow back
+// to the attacker's origin. Validate both RemoteAddr AND Host here.
+func (b *Broker) handleWebToken(w http.ResponseWriter, r *http.Request) {
+	if !isLoopbackRemote(r) || !hostHeaderIsLoopback(r) {
+		http.Error(w, "forbidden", http.StatusForbidden)
+		return
+	}
+	w.Header().Set("Content-Type", "application/json")
+	_ = json.NewEncoder(w).Encode(map[string]string{"token": b.token})
 }
